Refactored ServerRequest and ServerResponse with class protocols
